+++ xenon-common/src/main/java/com/vmware/xenon/common/UriUtils.java
@@ -61,6 +61,7 @@ public class UriUtils {
     public static final String URI_PARAM_ODATA_COUNT = "$count";
     public static final String URI_PARAM_ODATA_SKIP_TO = "$skipto";
     public static final String URI_PARAM_ODATA_NODE = "$nodeid";
+    public static final String URI_PARAM_ODATA_TENANTLINKS = "$tenantLinks";
     public static final String HTTP_SCHEME = "http";
     public static final String HTTPS_SCHEME = "https";
     public static final int HTTP_DEFAULT_PORT = 80;
@@ -719,6 +720,10 @@ public class UriUtils {
 
     public static String getODataNodeParamValue(URI uri) {
         return getODataParamValueAsString(uri, URI_PARAM_ODATA_NODE);
+    }
+
+    public static String getODataTenantLinksParamValue(URI uri) {
+        return getODataParamValueAsString(uri, URI_PARAM_ODATA_TENANTLINKS);
     }
 
     public static Integer getODataParamValue(final URI uri, final String uriParamOdataType) {