+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.9.3-SNAPSHOT
 
+* Support ODATA custom query options to specify tenantLinks. The
+  query param will be used to populate the tenantLinks field of
+  the underlying QueryTask
+
 * Rename TimeSeriesStats DataPoint data structure to TimeBin
 
 * Add support for groupBy queries. The new QueryOption.GROUP_BY, in